+++ mne_nirs/statistics/_glm_level_first.py
@@ -9,6 +9,7 @@ import warnings
 import pandas as pd
 import numpy as np
 from numpy import array_equal, where
+from h5io import read_hdf5, write_hdf5
 
 with warnings.catch_warnings(record=True):
     warnings.simplefilter('ignore')
@@ -19,7 +20,6 @@ from mne.channels.channels import ContainsMixin
 from mne.utils import fill_doc, warn, verbose, check_fname, _validate_type
 from mne.io.pick import _picks_to_idx
 from mne.io.constants import FIFF
-from mne.externals.h5io import read_hdf5, write_hdf5
 from mne import Info
 
 from ..visualisation._plot_GLM_topo import _plot_glm_topo,\
@@ -126,6 +126,8 @@ class _BaseGLM(ContainsMixin):
             Dataframe containing GLM results.
         """
         from ..utils import glm_to_tidy
+        if order is None:
+            order = self.ch_names
         return glm_to_tidy(self.info, self._data, self.design, order=order)
 
     def scatter(self, conditions=[], exclude_no_interest=True, axes=None,
@@ -374,12 +376,18 @@ class RegressionResults(_BaseGLM):
             integer indices of ``epochs.ch_names``). For example::
 
                 group_by=dict(Left_ROI=[1, 2, 3, 4], Right_ROI=[5, 6, 7, 8])
-
-            Note that within a dict entry all channels must have the same type.
         condition : str | list
             Name to be used for condition.
-        weighted : Bool
-            Should channels be weighted by inverse of standard error.
+        weighted : Bool | dict
+            Weighting to be applied to each channel in the ROI computation.
+            If False, then all channels will be weighted equally.
+            If True, channels will be weighted by the inverse of
+            the standard error of the GLM fit.
+            For manual specification of the channel weighting a dictionary
+            can be provided.
+            If a dictionary is provided, the keys and length of lists must
+            match the ``group_by`` parameters.
+            The weights will be scaled internally to sum to 1.
         demographic_info : Bool
             Add an extra column with demographic information from
             info["subject_info"].
@@ -392,6 +400,17 @@ class RegressionResults(_BaseGLM):
         if isinstance(condition, str):
             condition = [condition]
 
+        if isinstance(weighted, dict):
+            if weighted.keys() != group_by.keys():
+                raise KeyError("Keys of group_by and weighted "
+                               "must be the same")
+            for key in weighted.keys():
+                if len(weighted[key]) != len(group_by[key]):
+                    raise ValueError("The length of the keys for group_by "
+                                     "and weighted must match")
+                if (np.array(weighted[key]) < 0).any():
+                    raise ValueError("Weights must be positive values")
+
         tidy = pd.DataFrame()
         for cond in condition:
             cond_idx = where([c == cond for c in self.design.columns])[0]
@@ -399,6 +418,13 @@ class RegressionResults(_BaseGLM):
             roi = _glm_region_of_interest(self._data, group_by,
                                           cond_idx, cond, weighted)
             tidy = tidy.append(roi)
+
+        if weighted is True:
+            tidy["Weighted"] = "Inverse standard error"
+        elif weighted is False:
+            tidy["Weighted"] = "Equal"
+        elif isinstance(weighted, dict):
+            tidy["Weighted"] = "Custom"
 
         if demographic_info:
             if 'age' in self.info['subject_info'].keys():